requirements.txt
Django==4.0.4
pymongo==3.12.3
dnspython==2.2.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django_heroku==0.3.1
psycopg2==2.9.3
dj_database_url==0.5.0
whitenoise==6.2.0